Refactored redux elements
