+++ src/utils/auth/removeLocalStorageData.ts
@@ -1,0 +1,5 @@
+export const removeLocalStorageData = (): void => {
+	window.localStorage.removeItem("token");
+	window.localStorage.removeItem("username");
+	window.localStorage.removeItem("token-expiry");
+};
